requirements.txt
discord.py
aiohttp
python-dotenv
requests
browser-use
playwright
google-cloud-firestore
pandas
ccxt
notebooklm-py
websockets
